refactor: 인수테스트 픽스처 패키지 분리
