+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/MemberFixtures.java
@@ -1,4 +1,4 @@
-package com.woowacourse.gongseek.acceptance.support;
+package com.woowacourse.gongseek.acceptance.support.fixtures;
 
 import com.woowacourse.gongseek.auth.presentation.dto.AccessTokenResponse;
 import com.woowacourse.gongseek.member.presentation.dto.MemberUpdateRequest;
